+++ tests_integration/conftest.py
@@ -40,6 +40,7 @@ from tests.testing_utils.fixtures import (
 from tests.testing_utils.files_and_dirs import merge_left
 
 pytest_plugins = [
+    "tests_common",
     "tests_integration.testing_utils",
     "tests_integration.snowflake_connector",
 ]
